fd_readdir: makes dirent cache more like POSIX
This renames the type cached by multiple calls to fd_readdir to `Dir`
and its functions aligned with POSIX vs not aligned. This also adjusts
the errors they use to be consistent with what the underlying syscalls
return, e.g. ENOENT on invalid pos. Finally, this stops interpreting the
cookie as a signed number because the caller has no reason to interpret
it as a number and the logic is simpler if it isn't.

Signed-off-by: Adrian Cole <adrian@tetrate.io>